change: routes and controllers to different config"
