+++ app/controllers/application_controller.rb
@@ -1,5 +1,7 @@
 class ApplicationController < ActionController::API
 	rescue_from ActiveRecord::RecordNotFound, with: :error_response
+	rescue_from ActionController::ParameterMissing, with: :error_response
+
 
 	def error_response(exception)
 		error_serializer = ErrorSerializer.new(exception)
