+++ app/serializers/customer_subscription_serializer.rb
@@ -1,4 +1,4 @@
 class CustomerSubscriptionSerializer
   include JSONAPI::Serializer
-  attributes :customer_id, :subscription
+  attributes :customer_id, :status, :subscription
 end
