+++ config/routes.rb
@@ -4,9 +4,17 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   # root "articles#index"
 
+	# namespace :api do
+	# 	namespace :v1 do
+	# 		resources :customer_subscriptions, only: [:index, :create, :destroy, :update]
+	# 	end
+	# end
+
 	namespace :api do
 		namespace :v1 do
-			resources :customer_subscriptions, only: [:index, :create, :destroy, :update]
+			resources :customer, ony: :show do
+				resources :subscriptions, only: [:index, :create, :destroy, :update], module: 'customers'
+			end
 		end
 	end
 end
